Move replay data filling to training utils
